implementation/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
python-dotenv>=1.0.0
httpx>=0.24.0
scikit-learn>=1.2.2
lightgbm>=3.3.5
pandas>=2.0.0
numpy>=1.24.3
joblib>=1.2.0

# Data Analysis and Visualization
matplotlib>=3.7.1
seaborn>=0.12.2
scipy>=1.10.1

# Model Development
xgboost>=1.7.5
imbalanced-learn>=0.10.1

# Database
supabase>=1.0.3
asyncpg>=0.27.0

# Utilities
loguru>=0.7.0
colorama>=0.4.6
typer>=0.9.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Documentation
mkdocs>=1.4.3
mkdocs-material>=9.1.11 